refactor: remove ADR-0004 dead-code cascade (~1500 LOC)
ADR-0004 merged the old analyze command into scan, but batch.go and its
entire dependency cone were never deleted. This commit removes all of it.

Deleted files:
- internal/interfaces/cli/batch.go (1058 lines, 850 dead)
- internal/interfaces/cli/batch_test.go, batch_processing_test.go
- internal/interfaces/cli/option_predicates.go, option_predicates_test.go
- internal/interfaces/cli/purl_helpers_test.go (purlHasVersion etc.)
- internal/application/fetch_service.go (FetchService — zero live callers)
- internal/infrastructure/export/csv/license.go + license_test.go
- internal/common/result_processor.go + result_processor_test.go
- internal/common/collections/ (map_utils.go — only imported by batch.go)
- internal/common/purl/versionless.go + versionless_test.go

Extracted from batch.go into new input_helpers.go (live helpers scan.go uses):
- ProcessingOptions struct (slimmed: removed OnlyReviewNeeded, OnlyEOL,
  LicenseCSVPath, Ecosystem — no live readers after batch.go deletion)
- truncateDescription, randomSample, validateLineRange
- categorizeInputs, categorizeFileLines

Minimal edits to remaining files:
- application/analysis_service.go: removed WriteLicenseCSV
- common/errors.go: removed NewConfigError, NewInsufficientPermissionsError,
  IsInsufficientPermissionsError
- domain/config/value_objects.go: removed Service/Source/Validator interfaces,
  GetDefaultApp, GetDefaultDepsDev, deleted config field from ConfigService
- infrastructure/config/loader.go: removed GetConfig, Reload methods
- infrastructure/links/builder.go: removed three dead delegators
- domain/analysis/summary.go: removed FirstSentence, paragraphBreak, sentenceBreak
- domain/analysis/models.go: removed dead Score struct
- integration/populate_project.go: fixed stale comment referencing FirstSentence

All gates pass: go build ./..., go test ./..., go vet ./..., gofmt -l .
