+++ internal/domain/analysis/models_test.go
@@ -222,74 +222,6 @@ func TestVersionDetail(t *testing.T) {
 	}
 }
 
-func TestScore(t *testing.T) {
-	tests := []struct {
-		name  string
-		score Score
-		want  Score
-	}{
-		{
-			name: "valid_score",
-			score: Score{
-				Name:       "Security-Policy",
-				Value:      8,
-				Reason:     "Security policy is present",
-				Details:    []string{"Found SECURITY.md", "Contact info available"},
-				ComputedAt: time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC),
-			},
-			want: Score{
-				Name:       "Security-Policy",
-				Value:      8,
-				Reason:     "Security policy is present",
-				Details:    []string{"Found SECURITY.md", "Contact info available"},
-				ComputedAt: time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC),
-			},
-		},
-		{
-			name: "zero_score",
-			score: Score{
-				Name:       "Binary-Artifacts",
-				Value:      0,
-				Reason:     "Binary artifacts found",
-				Details:    []string{"Found: vendor/binary.exe"},
-				ComputedAt: time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC),
-			},
-			want: Score{
-				Name:       "Binary-Artifacts",
-				Value:      0,
-				Reason:     "Binary artifacts found",
-				Details:    []string{"Found: vendor/binary.exe"},
-				ComputedAt: time.Date(2023, 1, 1, 0, 0, 0, 0, time.UTC),
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.score.Name != tt.want.Name {
-				t.Errorf("Score.Name = %v, want %v", tt.score.Name, tt.want.Name)
-			}
-			if tt.score.Value != tt.want.Value {
-				t.Errorf("Score.Value = %v, want %v", tt.score.Value, tt.want.Value)
-			}
-			if tt.score.Reason != tt.want.Reason {
-				t.Errorf("Score.Reason = %v, want %v", tt.score.Reason, tt.want.Reason)
-			}
-			if len(tt.score.Details) != len(tt.want.Details) {
-				t.Errorf("Score.Details length = %v, want %v", len(tt.score.Details), len(tt.want.Details))
-			}
-			for i, detail := range tt.score.Details {
-				if detail != tt.want.Details[i] {
-					t.Errorf("Score.Details[%d] = %v, want %v", i, detail, tt.want.Details[i])
-				}
-			}
-			if !tt.score.ComputedAt.Equal(tt.want.ComputedAt) {
-				t.Errorf("Score.ComputedAt = %v, want %v", tt.score.ComputedAt, tt.want.ComputedAt)
-			}
-		})
-	}
-}
-
 func TestCommitStats(t *testing.T) {
 	tests := []struct {
 		name  string
